Fix warnings.

+++ src/tests.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2004, 2006, 2007 Free Software Foundation
+ * Copyright (C) 2004, 2006, 2007, 2008 Free Software Foundation
  * Copyright (C) 2000,2001,2002,2003 Nikos Mavrogiannopoulos
  *
  * This file is part of GNUTLS.
@@ -514,7 +514,7 @@ test_bye (gnutls_session_t session)
   old = siginterrupt (SIGALRM, 1);
   alarm (secs);
 #else
-  setsockopt (gnutls_transport_get_ptr (session), SOL_SOCKET, SO_RCVTIMEO,
+  setsockopt ((int)gnutls_transport_get_ptr (session), SOL_SOCKET, SO_RCVTIMEO,
 	      (char *) &secs, sizeof (int));
 #endif
